Remove global accelerator and use route53 instead.
This is due to a few issues around using AWS global accelerator (GA) for getting
a static IP to associate to the ELB created by EKS.
-  GA are only available for us-west-*
-  Terraform GA resource was buggy and was recreating the resource on every
apply.

When EKS supports k8s 1.15, Network Load balancers will have the ability
to support EIP allocations. However this is a more production like config,
so it's likely to be useful.